requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
PyYAML==6.0.1
joblib==1.3.2
